refactor(auth): extract implementation details from AuthenticationService
- Create AccessTokenUtil to handle JWT generation and verification logic
- Move email sending logic to PasswordResetTokenUtil.send_password_reset_email()
- Move token verification logic to PasswordResetTokenUtil.verify_password_reset_token()
- Clean AuthenticationService to show only public API methods
- Update test files to use new utility method locations
